찬성 반대 & 게임리스트 게임중 비활성화
