+++ middlewares/randomNickName.js
@@ -26,6 +26,11 @@ module.exports = async (req, res, next) => {
             '항해99 수료 한',
             '숙취로 괴로운',
             '여친 한테 차인',
+            '헛소리 하는',
+            '말이 너무 많은',
+            '입대 하루전',
+            '수능 답안지 밀려쓴',
+            '빛나는 대머리 의',
         ];
         const lastName = [
             '시고르자브종',
