+++ socket/sroom.js
@@ -1,230 +1,14 @@
 const { UserInfo, Room, User, Subject, Vote } = require('../models');
-const jwt = require('jsonwebtoken');
-const socketRandomName = require('../middlewares/socketRandomName');
+const {
+    makeRandomAvatar,
+    makeRandomNickName,
+    getRandomDebaters,
+    getRandomPosition,
+} = require('./randomFunc.js');
 const socketCheckLogin = require('../middlewares/socketCheckLogin');
-const socketRandomAvatar = require('../middlewares/socketRandomAvatar');
-const socketRandomNickName = require('../middlewares/socketRandomNickName.js');
 
 // // 중복 접속 방지
 // const connectedIPs = new Set();
-
-// 랜덤 아바타 생성함수
-const makeRandomAvatar = async (socket) => {
-    const avatarName = [
-        'Mary Baker',
-        'Amelia Earhart',
-        'Mary Roebling',
-        'Sarah Winnemucca',
-        'Margaret Brent',
-        'Lucy Stone',
-        'Mary Edwards',
-        'Margaret Chase',
-        'Mahalia Jackson',
-        'Maya Angelou',
-        'Margaret Bourke',
-        'Eunice Kennedy',
-        'Carrie Chapman',
-        'Elizabeth Peratrovich',
-        'Alicia Dickerson',
-        'Daisy Gatson',
-        'Emma Willard',
-        'Amelia Boynton',
-        'Maria Mitchell',
-        'Sojourner Truth',
-        'Willa Cather',
-        'Coretta Scott',
-        'Harriet Tubman',
-        'Fabiola Cabeza',
-        'Sacagawea',
-        'Esther Martinez',
-        'Elizabeth Cady',
-        'Bessie Coleman',
-        'Ma Rainey',
-        'Julia Ward',
-        'Irene Morgan',
-        'Babe Didrikson',
-        'Lyda Conley',
-        'Annie Dodge',
-        'Maud Nathan',
-        'Betty Ford',
-        'Rosa Parks',
-        'Susan La',
-        'Gertrude Stein',
-        'Wilma Mankiller',
-        'Grace Hopper',
-        'Jane Addams',
-        'Katharine Graham',
-        'Florence Chadwick',
-        'Zora Neale',
-        'Wilma Rudolph',
-        'Annie Jump',
-        'Mother Frances',
-        'Jovita Idár',
-        'Maggie L',
-        'Henrietta Swan',
-        'Jane Cunningham',
-        'Victoria Woodhull',
-        'Helen Keller',
-        'Patsy Takemoto',
-        'Chien-Shiung',
-        'Dorothea Dix',
-        'Margaret Sanger',
-        'Alice Paul',
-        'Frances Willard',
-        'Sally Ride',
-        'Juliette Gordon',
-        'Queen Lili',
-        'Katharine Lee',
-        'Harriet Beecher',
-        'Felisa Rincon',
-        'Hetty Green',
-        'Belva Lockwood',
-        'Biddy Mason',
-        'Ida B',
-        'Eleanor Roosevelt',
-        'Maria Goeppert',
-        'Phillis Wheatley',
-        'Mary Harris',
-        'Fannie Lou',
-        'Rosalyn Yalow',
-        'Susan B',
-        'Clara Barton',
-        'Lady Deborah',
-        'Jane Johnston',
-        'Alice Childress',
-        'Georgia O',
-        'Rebecca Crumpler',
-        'Anne Bradstreet',
-        'Elizabeth Blackwell',
-        'Christa McAuliffe',
-        'Edmonia Lewis',
-        'Nellie Bly',
-        'Mary Cassatt',
-        'Pauli Murray',
-        'Ellen Swallow',
-        'Hedy Lamarr',
-        'Pearl Kendrick',
-        'Abigail Adams',
-        'Margaret Fuller',
-        'Emma Lazarus',
-        'Marian Anderson',
-        'Virginia Apgar',
-        'Mary Walton',
-    ];
-
-    function randomRgb() {
-        let r = Math.floor(Math.random() * 256);
-        let g = Math.floor(Math.random() * 256);
-        let b = Math.floor(Math.random() * 256);
-        return [r, g, b];
-    }
-
-    function randomRgbHex() {
-        let [r, g, b] = randomRgb();
-        r = r.toString(16).padStart(2, '0');
-        g = g.toString(16).padStart(2, '0');
-        b = b.toString(16).padStart(2, '0');
-        return r + g + b;
-    }
-
-    console.log('****** 랜덤 아바타 함수시작 ******');
-    try {
-        const randomAvatarName =
-            avatarName[Math.floor(Math.random() * avatarName.length)];
-
-        const avatarColor = [];
-        while (avatarColor.length < 5) {
-            const colorHex = `#${randomRgbHex()}`;
-            avatarColor.push(colorHex);
-        }
-
-        const avatar = {
-            name: randomAvatarName,
-            color: avatarColor,
-        };
-
-        socket.avatar = avatar;
-    } catch (error) {
-        console.error('랜덤 아바타 생성에 실패했습니다:', error);
-        socket.emit('error', '랜덤 아바타 생성에 실패했습니다.');
-    }
-};
-
-const makeRandomNickName = async (socket) => {
-    try {
-        const firstName = [
-            '등짝스매싱 맞은',
-            '어제 과음 한',
-            '주식 대박 난',
-            '코인 쪽박 난',
-            '기타 치는',
-            '길가다 넘어 진',
-            '냄새나는',
-            '야동 보다 걸린',
-            '통장 잔고 700원',
-            '오디션 29104번 떨어진',
-            '청약 당첨 된',
-            '항해99 수료 한',
-            '숙취로 괴로운',
-            '여친 한테 차인',
-        ];
-        const lastName = [
-            '시고르자브종',
-            '웰시코기',
-            '도베르만',
-            '말티즈',
-            '보더콜리',
-            '개코원숭이',
-            '개미핥기',
-            '날다람쥐',
-            '아이언맨',
-            '배트맨',
-            '슈퍼맨',
-            '스파이더맨',
-            '우뢰매',
-            '황금박쥐',
-            '백터맨',
-            '동석이형',
-            '아메리카노',
-            '카푸치노',
-            '에스프레소',
-            '카라멜마끼야또',
-            '딸기주스',
-            '키위주스',
-            '오렌지',
-            '망고',
-            '토마토',
-            '애플망고',
-            '와플',
-            '또띠아',
-            '치즈버거',
-            '구찌',
-            '발렌시아가',
-            '샤넬',
-            '에르메스',
-            '페라리',
-            '람보르기니',
-            '제네시스',
-            '아우디',
-        ];
-
-        // 무작위로 firstName 배열 요소 선택
-        const randomAIndex = Math.floor(Math.random() * firstName.length);
-        const selectedA = firstName[randomAIndex];
-
-        // 무작위로 lastName 배열 요소 선택
-        const randomBIndex = Math.floor(Math.random() * lastName.length);
-        const selectedB = lastName[randomBIndex];
-
-        const nickName = selectedA + ' ' + selectedB;
-
-        socket.nickName = nickName;
-        console.log('****** 소켓 랜덤 닉네임 확인 완료 ******');
-    } catch (error) {
-        console.error('닉네임 부여 실패:', error);
-        socket.request.emit('error', '닉네임 부여에 실패했습니다.');
-    }
-};
 
 // 접속 유저 IP 체크
 const ipCheckFunc = async (socket) => {
@@ -255,7 +39,14 @@ const ipInfoDeleteFunc = async (socket) => {
 
 const getRoomList = async (kategorieId) => {
     const roomList = await Room.findAll({
-        attributes: ['roomId', 'KategorieName', 'roomName', 'debater', 'panel'],
+        attributes: [
+            'roomId',
+            'KategorieName',
+            'roomName',
+            'debater',
+            'panel',
+            'gameStart',
+        ],
         where: { kategorieId },
         order: [['createdAt', 'DESC']],
     });
@@ -297,10 +88,39 @@ async function updateRoomCount(roomId) {
     );
 }
 
+// Vote 테이블 data 삭제
+const deleteVotes = async () => {
+    try {
+        // Vote 테이블에서 모든 데이터 조회
+        const allVotes = await Vote.findAll();
+
+        // Room 테이블에서 모든 roomId 조회
+        const roomIds = await Room.findAll({
+            attributes: ['roomId'],
+        });
+
+        // Room 테이블에 존재하는 roomId 목록 생성
+        const existingRoomIds = roomIds.map((room) => room.roomId);
+
+        // Room 테이블에 존재하지 않는 Vote 데이터 필터링
+        const votesDelete = allVotes.filter(
+            (vote) => !existingRoomIds.includes(vote.roomId)
+        );
+
+        // Vote 테이블에서 삭제할 데이터 삭제
+        await Promise.all(votesDelete.map((vote) => vote.destroy()));
+
+        console.log('Vote 데이터를 정리하였습니다.');
+    } catch (error) {
+        console.error('Vote 데이터 정리에 실패하였습니다.', error);
+    }
+};
+
 module.exports = async (io) => {
     io.of('/roomList').on('connection', (socket) => {
         // 빈방 삭제
         deleteEmptyRooms();
+        deleteVotes();
 
         // // 접속 유저 IP 체크
         // ipCheckFunc();
@@ -495,6 +315,7 @@ module.exports = async (io) => {
 
                         // 빈방 삭제
                         await deleteEmptyRooms();
+                        await deleteVotes();
 
                         // 룸리스트 갱신
                         const roomList = await getRoomList(kategorieId);
@@ -513,14 +334,14 @@ module.exports = async (io) => {
                 // socket disconnecting
                 socket.on('disconnecting', async () => {
                     console.log('2. joinDebate disconnecting');
-                    // 나갈 유저정보 초기화
-                    userInfo.host = 0;
-                    userInfo.debater = 0;
                     userInfo.roomId = 0;
                     userInfo.like = 0;
                     userInfo.hate = 0;
+                    userInfo.debater = 0;
                     userInfo.questionMark = 0;
+                    userInfo.debaterPosition = 0;
                     await userInfo.save();
+                    // 나갈 유저정보 초기화
                 });
             } catch (error) {
                 console.error('토론자 참여 처리 실패:', error);
@@ -666,6 +487,7 @@ module.exports = async (io) => {
 
                         // 빈방 삭제
                         await deleteEmptyRooms();
+                        await deleteVotes();
 
                         const roomList = await getRoomList(kategorieId);
 
@@ -682,9 +504,25 @@ module.exports = async (io) => {
                 // socket disconnecting
                 socket.on('disconnecting', async () => {
                     console.log('3. joinJurror disconnecting');
-                    data = data.filter(
-                        (data) => data.userId !== userInfo.userId
-                    );
+                    userInfo.roomId = 0;
+                    userInfo.like = 0;
+                    userInfo.hate = 0;
+                    userInfo.questionMark = 0;
+                    await userInfo.save();
+
+                    // 방 인원 체크후 db업데이트
+                    await updateRoomCount(socket.roomId);
+
+                    // 빈방 삭제
+                    await deleteEmptyRooms();
+                    await deleteVotes();
+
+                    const roomList = await getRoomList(socket.kategorieId);
+
+                    // 네임스페이스에 룸리스트 보내기
+                    io.of('/roomList')
+                        .to(socket.kategorieId)
+                        .emit('update_roomList', roomList);
                 });
             } catch (error) {
                 console.error('배심원 참여 처리 실패:', error);
@@ -715,7 +553,7 @@ module.exports = async (io) => {
                 const allSubjects = JSON.parse(
                     subjectList.dataValues.subjectList
                 );
-                console.log('allSubjects=', allSubjects);
+                // console.log('allSubjects=', allSubjects);
 
                 const randomSubject = room.randomSubjects || []; // 기존 값이 NULL인 경우 빈 배열로 초기화
 
@@ -748,7 +586,7 @@ module.exports = async (io) => {
             return shuffled.slice(0, count); // 앞에서부터 count 개수만큼의 요소 반환
         }
 
-        socket.on('start_roulette', async (roomId, kategorieId, done) => {
+        socket.on('start_roulette', async (roomId, kategorieId) => {
             try {
                 const room = await Room.findOne({
                     where: { roomId },
@@ -773,7 +611,7 @@ module.exports = async (io) => {
                 // console.log('selectedSubject=', selectedSubject);
 
                 await Room.update(
-                    { roomName: selectedSubject },
+                    { roomName: selectedSubject, gameStart: 1 },
                     { where: { roomId } }
                 );
 
@@ -784,30 +622,71 @@ module.exports = async (io) => {
 
                 const roomList = await getRoomList(kategorieId);
 
-                io.of('/roomList')
+                await io
+                    .of('/roomList')
                     .to(kategorieId)
                     .emit('update_roomList', roomList);
+
                 io.to(roomId).emit('start_roulette', randomSubjectIndex);
-                done();
             } catch (error) {
                 console.error('룰렛 실행 실패:', error);
                 socket.emit('error', '룰렛 실행에 실패했습니다.');
             }
         });
 
-        socket.on('close_result', async (result, roomId, done) => {
+        socket.on('close_result', async (result, roomId) => {
             io.to(roomId).emit('close_result', result);
-            done();
         });
 
-        socket.on('close_roulette', async (result, roomId, done) => {
-            io.to(roomId).emit('close_roulette', result);
-            done();
-        });
+        socket.on(
+            'close_roulette',
+            async (result, roomId, kategorieId, done) => {
+                // 토론자 찬성, 반대 랜덤으로 선정
+                // UserInfo 테이블에서 roomId과 debater가 1인 사용자 목록 조회
+                const debaterList = await UserInfo.findAll({
+                    where: {
+                        roomId,
+                        debater: 1,
+                    },
+                });
+
+                // 랜덤으로 토론자 2명 선택
+                const randomDebaters = getRandomDebaters(debaterList, 2);
+
+                // 선택된 토론자에게 랜덤한 debatePosition 할당
+                const positions = [0, 1];
+                randomDebaters.forEach((debater) => {
+                    const randomPosition = getRandomPosition(positions);
+                    debater.debatePosition = randomPosition;
+                    debater.save();
+                });
+
+                // 게임상태 진행중으로 변경
+                await Room.update({ gameStart: 1 }, { where: { roomId } });
+                // 룸리스트 갱신
+                const roomList = await getRoomList(kategorieId);
+
+                // 네임스페이스에 룸리스트 보내기
+                io.of('/roomList')
+                    .to(kategorieId)
+                    .emit('update_roomList', roomList);
+
+                // 선택된 토론자의 정보를 프론트로 전달
+                const debatersInfo = randomDebaters.map((debater) => ({
+                    userId: debater.userId,
+                    host: debater.host,
+                    debatePosition: debater.debatePosition,
+                }));
+                console.log('프론트 전달 data', debatersInfo);
+
+                io.to(roomId).emit('close_roulette', result, debatersInfo);
+            }
+        );
 
         let debater1Count = 0;
         let debater2Count = 0;
-        socket.on('vote', async (roomId, host) => {
+        socket.on('vote', async (roomId, host, kategorieId) => {
+            console.log('투표 시 받아온 호스트 값', host);
             try {
                 // 룸 정의
                 const room = await Room.findOne({
@@ -837,14 +716,19 @@ module.exports = async (io) => {
                         debater2Count: 0,
                     });
                 }
+                console.log('순서대로 진행되는 지 확인');
                 // 전달받은 host값이 1이면 1번토론자 투표수 증가 , 0이면 2번토론자 투표수 증가
                 if (host === 1) {
+                    console.log('호스트 투표');
                     await voteRecord.increment('debater1Count');
                 } else if (host === 0) {
+                    console.log('도전자 투표');
                     await voteRecord.increment('debater2Count');
                 }
+
                 // 투표 수 가져오기
                 await voteRecord.reload();
+                console.log(voteRecord);
                 const voteCount =
                     voteRecord.debater1Count + voteRecord.debater2Count;
                 console.log('투표수', voteCount);
@@ -854,21 +738,68 @@ module.exports = async (io) => {
                     let loser;
                     let winnerCount;
                     let loserCount;
-                    if (voteRecord.debater1Count > voteRecord.debater2Count) {
+                    if (voteRecord.debater1Count === voteRecord.debater2Count) {
+                        console.log('뭐여 무승부여?');
+                        const voteResult = {
+                            debater1: debaterUser1.userId,
+                            debater1Count: voteRecord.debater1Count,
+                            debater2: debaterUser2.userId,
+                            debater2Count: voteRecord.debater2Count,
+                        };
+
+                        await io
+                            .to(roomId)
+                            .emit('voteResult', voteResult, () => {
+                                io.to(roomId).emit('roomJoined', data);
+                            });
+
+                        voteRecord.debater1Count = 0;
+                        voteRecord.debater2Count = 0;
+                        await voteRecord.save();
+
+                        console.log(
+                            '(무승부)투표수 초기화',
+                            voteRecord.debater1Count
+                        );
+                        console.log(
+                            '(무승부)투표수 초기화',
+                            voteRecord.debater2Count
+                        );
+
+                        // 투표 종료 후 room 상태 게임종료 로 업데이트
+                        await Room.update(
+                            { gameStart: 0 },
+                            { where: { roomId } }
+                        );
+
+                        const roomList = await getRoomList(kategorieId);
+
+                        // 네임스페이스에 룸리스트 보내기
+                        io.of('/roomList')
+                            .to(kategorieId)
+                            .emit('update_roomList', roomList);
+
+                        return;
+                    } else if (
+                        voteRecord.debater1Count > voteRecord.debater2Count
+                    ) {
+                        console.log('방장이 이겼네');
                         winner = debaterUser1;
                         winnerCount = voteRecord.debater1Count;
                         loser = debaterUser2;
                         loserCount = voteRecord.debater2Count;
                     } else {
+                        console.log('도전자가 이겼네');
                         winner = debaterUser2;
                         winnerCount = voteRecord.debater2Count;
                         loser = debaterUser1;
                         loserCount = voteRecord.debater1Count;
                     }
                     const voteResult = {
-                        winner: winner.nickName,
+                        winner: winner.userId,
+                        winnerNickName: winner.nickName,
                         winnerCount: winnerCount,
-                        loser: loser.nickName,
+                        loser: loser.userId,
                         loserCount: loserCount,
                     };
                     console.log(
@@ -876,49 +807,26 @@ module.exports = async (io) => {
                         `우승자는 ${winnerCount}표를 받은 ${winner.nickName}입니다. 패자는 ${loserCount}표를 받은 ${loser.nickName}입니다.`
                     );
                     console.log('voteResult', voteResult);
-                    io.to(roomId).emit('voteResult', voteResult);
+                    await io.to(roomId).emit('voteResult', voteResult);
                     // 투표 종료 후 데이터 보내주고 voteCount 초기화
                     voteRecord.debater1Count = 0;
                     voteRecord.debater2Count = 0;
                     await voteRecord.save();
-                    console.log('투표수 초기화', voteRecord.debater1Count);
-                    console.log('투표수 초기화', voteRecord.debater2Count);
-                    // winner 정보 초기화
-                    // await UserInfo.update(
-                    //     {
-                    //         like: 0,
-                    //         hate: 0,
-                    //         questionMark: 0,
-                    //         debater: 1,
-                    //         host: 1,
-                    //     },
-                    //     { where: { userId: winner.userId } }
-                    // );
-                    // // loser 정보 초기화
-                    // await UserInfo.update(
-                    //     {
-                    //         like: 0,
-                    //         hate: 0,
-                    //         questionMark: 0,
-                    //         debater: 0,
-                    //         host: 0,
-                    //         roomId: 0,
-                    //     },
-                    //     { where: { userId: loser.userId } }
-                    // );
+
+                    // 투표 종료 후 room 상태 게임종료 로 업데이트
+                    await Room.update({ gameStart: 0 }, { where: { roomId } });
+
+                    console.log(
+                        '(결과나옴)투표수 초기화',
+                        voteRecord.debater1Count
+                    );
+                    console.log(
+                        '(결과나옴)투표수 초기화',
+                        voteRecord.debater2Count
+                    );
                     console.log('승자 닉네임', winner.nickName);
                     console.log('승자 디베이터', winner.debater);
                     console.log('데이터 =', data);
-                    // data = data.map((item) =>
-                    //     item.userId === winner.userId
-                    //         ? { ...item, host: 1, debater: 1 }
-                    //         : item
-                    // );
-                    // data = data.map((item) =>
-                    //     item.userId === loser.userId
-                    //         ? { ...item, host: 0, debater: 0 }
-                    //         : item
-                    // );
                     console.log('투표후', data);
                     console.log('패배자Id', loser.userId);
                     // loser 퇴장 시키기
@@ -960,45 +868,58 @@ module.exports = async (io) => {
 
         socket.on('disconnecting', async () => {
             console.log('1. connection disconnecting');
-            const exitUserInfo = await UserInfo.findOne({
-                where: { userId: socket.user.userId },
-            });
-
-            console.log('1. 종료한 유저의 호스트 값 = ', exitUserInfo.host);
-
-            if (exitUserInfo.host === 1) {
-                const newHost = await UserInfo.findOne({
-                    where: {
-                        roomId: socket.roomId,
-                        host: 0,
-                        debater: 1,
-                    },
+            if (socket.user) {
+                const exitUserInfo = await UserInfo.findOne({
+                    where: { userId: socket.user.userId },
                 });
-                if (newHost) {
-                    newHost.host = 1;
-                    await newHost.save();
-                    data.map((data) => {
-                        if (data.userId === newHost.userId) {
-                            data.host = 1;
-                        }
+
+                console.log('1. 종료한 유저의 호스트 값 = ', exitUserInfo.host);
+
+                if (exitUserInfo.host === 1) {
+                    const newHost = await UserInfo.findOne({
+                        where: {
+                            roomId: socket.roomId,
+                            host: 0,
+                            debater: 1,
+                        },
                     });
-                    io.to(socket.roomId).emit('changeHost', newHost.userId);
+                    if (newHost) {
+                        newHost.host = 1;
+                        await newHost.save();
+                        data.map((data) => {
+                            if (data.userId === newHost.userId) {
+                                data.host = 1;
+                            }
+                        });
+                        io.to(socket.roomId).emit('changeHost', newHost.userId);
+                    }
+                    exitUserInfo.host = 0;
+                    await exitUserInfo.save();
+                }
+
+                data = data.filter(
+                    (item) => item.userId !== socket.user.userId
+                );
+
+                // 방 퇴장 후 남아있는 userData 보내기
+                if (socket.roomId) {
+                    io.to(socket.roomId).emit('roomJoined', data);
+
+                    // 방 인원 체크후 db업데이트
+                    await updateRoomCount(socket.roomId);
+                }
+                // 빈방 삭제
+                await deleteEmptyRooms();
+
+                if (socket.kategorieId) {
+                    const roomList = await getRoomList(socket.kategorieId);
+
+                    // 네임스페이스에 룸리스트 보내기
+                    io.of('/roomList')
+                        .to(socket.kategorieId)
+                        .emit('update_roomList', roomList);
                 }
             }
-            data = data.filter((item) => item.userId !== socket.user.userId);
-
-            // 방 퇴장 후 남아있는 userData 보내기
-            io.to(socket.roomId).emit('roomJoined', data);
-
-            // 방 인원 체크후 db업데이트
-            await updateRoomCount(socket.roomId);
-
-            const roomList = await getRoomList(socket.kategorieId);
-
-            // 네임스페이스에 룸리스트 보내기
-            io.of('/roomList')
-                .to(socket.kategorieId)
-                .emit('update_roomList', roomList);
         });
     });
 };
